% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart-utils.R
\name{as.character.patient_flowchart}
\alias{as.character.patient_flowchart}
\title{Text representation of patient inclusion flowchart}
\usage{
\method{as.character}{patient_flowchart}(x, length = 7, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{length}{Length of the arrows (to the right)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Text representation of patient inclusion flowchart
}
