% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{assign_list}
\alias{assign_list}
\title{Assign variables in a list}
\usage{
assign_list(x, envir = .GlobalEnv)
}
\arguments{
\item{x}{A named list of values}

\item{envir}{The environment in which the values are assigned, defaults to the global environment}
}
\description{
Assign variables in a list
}
\examples{
assign_list(list(a=1, b=2))
}
