% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-utils.R
\name{create_text_table}
\alias{create_text_table}
\title{Creates a text table}
\usage{
create_text_table(string, table_width = 80, compact = TRUE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{table_width}{table character width.}

\item{compact}{whether to take only the necessary space (\code{TRUE}) or to fill out the table_width (\code{FALSE}).}
}
\value{
A vector of strings per row, forming together a table.
}
\description{
Creates a text table
}
\examples{
cat(create_text_table(LETTERS),sep = "\n")
}
\seealso{
\code{\link{get_square_grid}}.
}
