% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{default}
\alias{default}
\title{Default value}
\usage{
default(x, default_value)
}
\arguments{
\item{x}{A value}

\item{default_value}{The replacement value for when x is NA, NULL or has a length of 0.}
}
\value{
The value of x when x is not NA, NULL or has a length of 0, and default_value otherwise.
}
\description{
Returns a default value for a scalar, to be used when the input is NA, NULL or has a length of 0.
}
\examples{
default(NA, 0)
}
