% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{.get_title_bar}
\alias{.get_title_bar}
\title{Creates a title bar}
\usage{
.get_title_bar(left = NULL)
}
\arguments{
\item{left}{The text on the left side of the title bar, may be \code{NULL}}
}
\value{
A string vector of a title bar of 80 characters.
}
\description{
Creates a title bar
}
\keyword{internal}
