% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-utils.R
\name{.pkg_duplicated}
\alias{.pkg_duplicated}
\title{Find duplicated packages names}
\usage{
.pkg_duplicated(pkgs)
}
\arguments{
\item{pkgs}{A list of packages names}
}
\value{
A named list of duplicated names and number of occurrences
}
\description{
Find duplicated packages names
}
