% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-utils.R
\name{.regexl}
\alias{.regexl}
\title{Extracts the matches from stringr::str_match[_all]}
\usage{
.regexl(result)
}
\arguments{
\item{result}{The results from stringr::str_match[_all]}
}
\value{
a list of matches
}
\description{
Extracts the matches from stringr::str_match[_all]
}
