% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_square_grid}
\alias{get_square_grid}
\title{Specifies a square grid which fits N objects.}
\usage{
get_square_grid(N, moreRows = TRUE)
}
\arguments{
\item{N}{number of objects.}

\item{moreRows}{whether there should be more rows than columns if the resulting grid is not square. Defaults to more rows (\code{TRUE}).}
}
\value{
A named list with elements rows and columns specifying the size of the optimal grid.
}
\description{
The resulting grid will be of size \code{a*a} or \code{a*(a+1)} where \code{a} is an integer.
It will therefore always be a square or or have one row/column more than columns/rows.
}
\examples{
get_square_grid(5)
}
