% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-loading.R
\name{load_package_collection}
\alias{load_package_collection}
\alias{list_package_collections}
\alias{list_common_packages}
\alias{load_common_packages}
\title{List package collections}
\usage{
load_package_collection(
  collection_name = names(list_package_collections()),
  ...
)

list_package_collections()

list_common_packages()

load_common_packages(...)
}
\arguments{
\item{collection_name}{One or multiple collection names. Must be in \code{"data_import","image_import","ggplot",
"grid","survival","processing","shiny","development"}.}

\item{...}{list of package names.}
}
\description{
List package collections
}
