% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-utils.R
\name{messages}
\alias{messages}
\alias{mess_info}
\alias{mess_succ}
\alias{mess_fail}
\alias{mess_warn}
\alias{mess_numb}
\alias{mess_hint}
\title{Display coloured messages}
\usage{
mess_info(header, message, pre = "")

mess_succ(header, message, pre = "")

mess_fail(header, message, pre = "")

mess_warn(header, message, pre = "")

mess_numb(header, message, pre = "")

mess_hint(message, header = "", pre = "")
}
\arguments{
\item{header}{message title.}

\item{message}{the message.}

\item{pre}{optional string to be displayed before the message bullet (such as \code{'\n'}).}
}
\value{
A coloured message.
}
\description{
Display coloured messages
}
\keyword{internal}
