% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart-utils.R
\name{print.patient_flowchart}
\alias{print.patient_flowchart}
\title{Print the patient inclusion flowchart}
\usage{
\method{print}{patient_flowchart}(x, length = 7, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{length}{Length of the arrows (to the right)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print the patient inclusion flowchart
}
