% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-utils.R
\name{redundant_packages}
\alias{redundant_packages}
\title{Find redundant packages}
\usage{
redundant_packages(packages)
}
\arguments{
\item{packages}{list of package names.}
}
\value{
A named list of packages names, where each value is a vector of packages already loading the corresponding package.
}
\description{
Find redundant packages
}
\details{
Certain packages have a direct dependency on other packages. In that case it is unnecessary to attach the latter packages.
This function finds those packages and returns them in a named list. For each named item, the name is imported by the value in the list.
}
\examples{
\dontrun{
#grid does not have be loaded since gridGraphics already does so.
redundant_packages(c("gridGraphics","grid"))
}
}
