% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{rm_empty_rows}
\alias{rm_empty_rows}
\title{Remove empty rows}
\usage{
rm_empty_rows(dataframe)
}
\arguments{
\item{dataframe}{\code{data.frame} object.}
}
\value{
A \code{data.frame} with rows removed that only contain \code{NA}.
}
\description{
Remove empty rows
}
\examples{
data <- rbind(c(1,2,3), c(1, NA, 4), c(4,6,7), c(NA, NA, NA), c(4, 8, NA))
rm_empty_rows(data)
}
\seealso{
Other NA functions: 
\code{\link{rm_na}()}
}
\concept{NA functions}
