% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{rm_na}
\alias{rm_na}
\title{Remove \code{NA}}
\usage{
rm_na(x)
}
\arguments{
\item{x}{vector containing possible \code{NA} values.}
}
\value{
Vector without \code{NA}
}
\description{
Remove \code{NA}
}
\examples{
rm_na(c(1,2,NA,54))
}
\seealso{
Other NA functions: 
\code{\link{rm_empty_rows}()}
}
\concept{NA functions}
