% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{rnd_dbl}
\alias{rnd_dbl}
\title{Round number}
\usage{
rnd_dbl(dbl, digits = 3)
}
\arguments{
\item{dbl}{number to be rounded.}

\item{digits}{number of digits the number needs to be rounded to (defaults to \code{3}).}
}
\value{
A string value of the number rounded to the specified amount of digits.
}
\description{
Rounds a number to a specified amount of digits and returns the string value.
}
\examples{
rnd_dbl(1.26564,digits = 2)
}
