% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{sep_thousands}
\alias{sep_thousands}
\title{Adds comma's to separate thousands in numbers}
\usage{
sep_thousands(n)
}
\arguments{
\item{n}{a real number}
}
\value{
A string with the number and thousands separated by comma's.
}
\description{
Adds comma's to separate thousands in numbers
}
\examples{
sep_thousands(13243.33) #13,243.33
}
