% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{startup}
\alias{startup}
\title{Cleans R for use}
\usage{
startup(
  removeObjects = TRUE,
  runGarbageCollection = TRUE,
  clearGraphics = TRUE,
  folder = NULL,
  verbose = TRUE,
  seed = 37
)
}
\arguments{
\item{removeObjects}{whether to remove objects from the workspace.}

\item{runGarbageCollection}{whether to run the garbage collection.}

\item{clearGraphics}{whether to clear the graphics from the R studio plots screen.}

\item{folder}{folder name to set the current working directory.}

\item{verbose}{whether to print informative messages during cleaning.}

\item{seed}{the random set to be set with set.seed; is ignored if value is set to NULL.}
}
\description{
Clears workspace, deletes all objects from global environment,
clears graphics and (optionally) sets working directory.
}
\examples{
\dontrun{startup()}
}
\concept{initialization functions}
