% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{translate_items}
\alias{translate_items}
\title{Translate item}
\usage{
translate_items(vector, dict)
}
\arguments{
\item{vector}{A vector whose values are to be translated.}

\item{dict}{A named vector, whose names are keys in 'vector' to be replaced and whose values are the new values}
}
\value{
A vector with new values
}
\description{
Translate item
}
\examples{
v = c("A","B","C")
dict = c("A"="1")

translate_items(v, dict)
}
